+++ examples/nn_dnn_test.py
@@ -77,7 +77,7 @@ def _main():
                                name=model_name,
                                **vars(args)
                                )
-    model = model.to(dv)
+    model = model.to(dv).double()
 
     if args.mode == 'train':
         model.fit(train_dataset_producer,
